Add comment support and a reserved hierarchy type
Support parsing multi-line vectors in the configuration syntax.

Add a comment field to the configuration syntax and include its value
in control.ctl before the metadata headers.

Add a new reserved hierarchy type, for hierarchies that are reserved
for special purposes.  This is slightly different than any of the
existing statuses.