Support chkscope (partially) in checkgroups messages
Add support for chkscope parameters in the checkgroups control
message.  process-control now correctly filters out out-of-scope lines
and doesn't apply the checkgroups to out-of-scope entries in the
current active file, but the scope is not yet correctly used to
determine the matching control.ctl entry.  Should fix problems with
processing ka.* checkgroups.